_legacy_2026-02-07T20:30:00/requirements.txt
PySide6==6.8.0
SQLAlchemy==2.0.36
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
dropbox>=11.0.0
